# use memory pool for storeDicts