--- a/builder/generated_static_code.go
+++ b/builder/generated_static_code.go
@@ -706,7 +706,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {