--- a/examples/calculator/calculator.go
+++ b/examples/calculator/calculator.go
@@ -1126,7 +1126,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {