--- a/test/andnot/andnot.go
+++ b/test/andnot/andnot.go
@@ -841,7 +841,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {