--- a/test/errorpos/errorpos.go
+++ b/test/errorpos/errorpos.go
@@ -1256,7 +1256,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {