--- a/test/global_store/global_store.go
+++ b/test/global_store/global_store.go
@@ -859,7 +859,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {