--- a/test/goto_state/goto_state.go
+++ b/test/goto_state/goto_state.go
@@ -1104,7 +1104,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {