--- a/test/issue_1/issue_1.go
+++ b/test/issue_1/issue_1.go
@@ -780,7 +780,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {