--- a/test/issue_18/issue_18.go
+++ b/test/issue_18/issue_18.go
@@ -798,7 +798,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {