--- a/test/issue_65/issue_65.go
+++ b/test/issue_65/issue_65.go
@@ -794,7 +794,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {