--- a/test/issue_70/issue_70.go
+++ b/test/issue_70/issue_70.go
@@ -771,7 +771,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {