--- a/test/issue_70/optimized-grammar/issue_70.go
+++ b/test/issue_70/optimized-grammar/issue_70.go
@@ -755,7 +755,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {