--- a/test/issue_80/issue_80.go
+++ b/test/issue_80/issue_80.go
@@ -791,7 +791,7 @@ type Cloner interface {
 }
 
 var statePool = &sync.Pool{
-	New: func() interface{} { return make(storeDict, 2) },
+	New: func() interface{} { return make(storeDict) },
 }
 
 func (sd storeDict) Discard() {