This is important to reduce GC pressure in applications that spend a lot of time parsing. Let me know if you want more details on this, but we're already using this on our end, so would be great if we can merge it upstream.

@mna @breml 

EDIT: also let me know if you think we should write benchmarks to verify this