@breml 2019-07-16
@metalogical Thanks for your PR. The PR looks pretty straight forward to me and it makes sense, that if a long running process uses the parse a lot, there is some pressure on the GC. I would really appreciate, if you could provide some benchmarks, that show the positive effect of the `sync.Pool` in that case. Additionally it would be interesting to investigate, if there are other spots, where maps are used, that could benefit from `sync.Pool` as well.