@hrysoula 2019-07-17
> Please run make clean && make all. The file bootstrap/cmd/bootstrap-pigeon/bootstrap_pigeon.go is not up to date in the last commit.

Fixed in e25b068.