@hrysoula 2019-08-06
Internally we have run benchmarks that verify the improvement. I unfortunately cannot provide the parser as it is a private repository.

Before:
```
goos: linux
goarch: amd64
pkg: github.com/kiteco/kiteco/kite-go/lang/python/pythonparser/calls/internal/pigeon
BenchmarkValidStrings-4     	     500	   3367479 ns/op
BenchmarkInvalidStrings-4   	   10000	    180598 ns/op
PASS
```

After:
```
goos: linux
goarch: amd64
pkg: github.com/kiteco/kiteco/kite-go/lang/python/pythonparser/calls/internal/pigeon
BenchmarkValidStrings-4     	     500	   2808122 ns/op
BenchmarkInvalidStrings-4   	   10000	    131682 ns/op
PASS
```

Running the existing [`benchparse_test.go`](https://github.com/mna/pigeon/blob/master/benchparse_test.go) tests, the improvement is less noticeable but still present:

Before:
```
goos: darwin
goarch: amd64
pkg: github.com/mna/pigeon
BenchmarkParseUnicodeClass-8   	   10000	    216224 ns/op
BenchmarkParseKeyword-8        	   10000	    135896 ns/op
PASS
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/mna/pigeon
BenchmarkParseUnicodeClass-8   	   10000	    212132 ns/op
BenchmarkParseKeyword-8        	   10000	    135354 ns/op
PASS
```