@metalogical 2019-08-09
@breml are the existing benchmarks that Hrysoula posted above enough, or would you like something more to get this merged?

cc @mna