@breml 2019-08-10
@hrysoula Thanks for the benchmarks. These look good to me. So I will merge this.